Using double for coordinates
git-svn-id: file:///private-backup/var/svn/jngl@19 8b392cba-0c28-4fbc-ad53-5d9c93102ef3

+++ window.cpp
@@ -22,7 +22,7 @@ along with jngl.  If not, see <http://www.gnu.org/licenses/>.
 
 namespace jngl
 {
-	void Window::Print(const std::string& text, const int xposition, const int yposition)
+	void Window::Print(const std::string& text, const double xposition, const double yposition)
 	{
 		fonts_[fontSize_][fontName_].Print(xposition, yposition, text);
 	}
